Extract withServer and export Fields test to a separate file
